requirements.txt
google-generativeai
python-dotenv
PyMuPDF
pandas
tqdm
